requirements.txt
flask
rembg
python-dotenv
Pillow